DNS/pyproject.toml
[tool.poetry]
name = "dns"
version = "0.1.0"
description = ""
authors = ["Bogdan <bogdan.vtk.4516@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
dnspython = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
